- Add object creation - FIX Brightness Action - FIX Cors error

+++ src/main/java/app/patronuscontrol/model/action/Brightness.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Brightness extends Action {
 
-    private float brightness;
+    private int brightness;
 
     @Override
     public Attribute getAttribute() {
